backend/microservices/requirements_whisper.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai-whisper==20231117
torch==2.2.2
torchaudio==2.2.2
ffmpeg-python==0.2.0
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0
